feat: #103, #76 작업 완료
- 타워에 경우 현재 구현한 타워만 설치 가능하도록 구현
- 행성의 경우 패시브를 제외한 스탯만 가져와서 사용할 수 있도록 구현
- 프리셋 기능 완성

+++ Assets/Script/DefenceTower/RandomOption/TowerDamageUpgradeOption.cs
@@ -13,7 +13,7 @@ public class TowerDamageUpgradeOption : RandomOptionBase
     {
         foreach(var tower in towers)
         {
-            tower.AddBonusDamage(-baseTowerData.optionValue);
+            tower.AddBonusDamage((int)-baseTowerData.optionValue);
         }
     }
 
@@ -25,7 +25,7 @@ public class TowerDamageUpgradeOption : RandomOptionBase
 #if DEBUG_MODE
             //Debug.Log($"Before Damage Amount Apply : {tower.Damage} towerID : {tower.ID}");
 #endif
-            tower.AddBonusDamage(baseTowerData.optionValue);
+            tower.AddBonusDamage((int)baseTowerData.optionValue);
 #if DEBUG_MODE
             //Debug.Log($"After Damage Amount Apply : {tower.Damage} towerID : {tower.ID}");
 #endif
